0.0.b57

+++ evaluation_framework/evaluation_engine.py
@@ -23,28 +23,8 @@ from collections import namedtuple
 import psutil
 import shutil
 import time
+import copy
 
-
-# INSTANCE_TYPES = {
-#     'm4.large': {'vCPU': 2, 'Mem': 8},
-#     'm4.xlarge': {'vCPU': 4, 'Mem': 16}, 
-#     'm4.2xlarge': {'vCPU': 8, 'Mem': 32},
-#     'm4.4xlarge': {'vCPU': 16, 'Mem': 64},
-#     'm4.10xlarge': {'vCPU': 40, 'Mem': 160},
-#     'm4.16xlarge': {'vCPU': 64, 'Mem': 256}, 
-    
-#     'c4.large': {'vCPU': 2, 'Mem': 3.75},
-#     'c4.xlarge': {'vCPU': 4, 'Mem': 7.5},
-#     'c4.2xlarge': {'vCPU': 8, 'Mem': 15},
-#     'c4.4xlarge': {'vCPU': 16, 'Mem': 30},
-#     'c4.8xlarge': {'vCPU': 36, 'Mem': 60}, 
-    
-#     'r4.large': {'vCPU': 2, 'Mem': 15.25},
-#     'r4.xlarge': {'vCPU': 4, 'Mem': 30.5},
-#     'r4.2xlarge': {'vCPU': 8, 'Mem': 61}, 
-#     'r4.4xlarge': {'vCPU': 16, 'Mem': 122},
-#     'r4.8xlarge': {'vCPU': 32, 'Mem': 244},
-#     'r4.16xlarge': {'vCPU': 64, 'Mem': 488}}
 
 INSTANCE_TYPES = {
     'm4.large': {'vCPU': 2, 'Mem': 8},
@@ -60,12 +40,33 @@ INSTANCE_TYPES = {
     'c4.4xlarge': {'vCPU': 16, 'Mem': 30},
     'c4.8xlarge': {'vCPU': 36, 'Mem': 60}, 
     
-    'r4.large': {'vCPU': 2, 'Mem': 15.25 - 3},
-    'r4.xlarge': {'vCPU': 4, 'Mem': 30.5 - 6},
-    'r4.2xlarge': {'vCPU': 8, 'Mem': 61 - 12}, 
-    'r4.4xlarge': {'vCPU': 16, 'Mem': 122 - 25},
-    'r4.8xlarge': {'vCPU': 32, 'Mem': 244 - 50},
-    'r4.16xlarge': {'vCPU': 64, 'Mem': 488 - 50}}
+    'r4.large': {'vCPU': 2, 'Mem': 15.25},
+    'r4.xlarge': {'vCPU': 4, 'Mem': 30.5},
+    'r4.2xlarge': {'vCPU': 8, 'Mem': 61}, 
+    'r4.4xlarge': {'vCPU': 16, 'Mem': 122},
+    'r4.8xlarge': {'vCPU': 32, 'Mem': 244},
+    'r4.16xlarge': {'vCPU': 64, 'Mem': 488}}
+
+# INSTANCE_TYPES = {
+#     'm4.large': {'vCPU': 2, 'Mem': 8},
+#     'm4.xlarge': {'vCPU': 4, 'Mem': 16}, 
+#     'm4.2xlarge': {'vCPU': 8, 'Mem': 32},
+#     'm4.4xlarge': {'vCPU': 16, 'Mem': 64},
+#     'm4.10xlarge': {'vCPU': 40, 'Mem': 160},
+#     'm4.16xlarge': {'vCPU': 64, 'Mem': 256}, 
+    
+#     'c4.large': {'vCPU': 2, 'Mem': 3.75},
+#     'c4.xlarge': {'vCPU': 4, 'Mem': 7.5},
+#     'c4.2xlarge': {'vCPU': 8, 'Mem': 15},
+#     'c4.4xlarge': {'vCPU': 16, 'Mem': 30},
+#     'c4.8xlarge': {'vCPU': 36, 'Mem': 60}, 
+    
+#     'r4.large': {'vCPU': 2, 'Mem': 15.25 - 3},
+#     'r4.xlarge': {'vCPU': 4, 'Mem': 30.5 - 6},
+#     'r4.2xlarge': {'vCPU': 8, 'Mem': 61 - 12}, 
+#     'r4.4xlarge': {'vCPU': 16, 'Mem': 122 - 25},
+#     'r4.8xlarge': {'vCPU': 32, 'Mem': 244 - 50},
+#     'r4.16xlarge': {'vCPU': 64, 'Mem': 488 - 50}}
 
 DEFAULT_LARGE_INSTANCE_WORKER_VCORES = 4
 DEFAULT_SMALL_INSTANCE_WORKER_VCORES = 2
@@ -124,6 +125,7 @@ class EvaluationEngine():
             n_worker_nodes, use_yarn_cluster, use_auto_config, instance_type)
 
         self.has_dask_client = False
+        self.has_data = False
         
     def run_evaluation(self, evaluation_manager, debug_mode=False):
 
@@ -132,16 +134,25 @@ class EvaluationEngine():
         if self.use_yarn_cluster and evaluation_manager.S3_path is None:
             raise ValueError('if [ use_yarn_cluster ] is set to True, you must provide [ S3_path ] to EvaluationManager object.')
 
-        if os.path.exists(evaluation_manager.evaluation_task_dirpath):
-            print('\u2757 Removing duplicate evaluation_task_dirpath\n')
-            shutil.rmtree(evaluation_manager.evaluation_task_dirpath)
-        os.makedirs(evaluation_manager.evaluation_task_dirpath)
+        if not evaluation_manager.local_data_saved:
+
+            if os.path.exists(evaluation_manager.evaluation_task_dirpath):
+                print('\u2757 Removing duplicate evaluation_task_dirpath\n')
+                shutil.rmtree(evaluation_manager.evaluation_task_dirpath)
+            os.makedirs(evaluation_manager.evaluation_task_dirpath)
+
+        else:
+
+            print('\u2757 Reusing previous evaluation data and dask cluster')
+
+            self.taskq.flush_results()
+        
         os.chdir(evaluation_manager.evaluation_task_dirpath)
-        # by not removing the local_directory_path (root) but just the task specific dir, 
-        # we can ensure re-runnability of the current evaluation task.
-        # if EM were to be redefined, another task dir will be created.
-        # the change of directory is required for sharing methods across yarn and local clients
-        # also, need to start dask AFTER the change in directory 
+            # by not removing the local_directory_path (root) but just the task specific dir, 
+            # we can ensure re-runnability of the current evaluation task.
+            # if EM were to be redefined, another task dir will be created.
+            # the change of directory is required for sharing methods across yarn and local clients
+            # also, need to start dask AFTER the change in directory 
 
 
         if(not debug_mode):
@@ -150,15 +161,19 @@ class EvaluationEngine():
                 self.start_dask_client()
                 self.has_dask_client = True
             else:
-                self.stop_dask_client()
-                self.start_dask_client()
-                self.has_dask_client = True
-                
-        print("\u2714 Preparing local data...            ", end="", flush=True)
-        print()
-        # self.memmap_map = load_local_data(evaluation_manager)
-        load_local_data(evaluation_manager)
-        # print('Completed!')
+                # reuse the client
+                pass
+                # self.stop_dask_client()
+                # self.start_dask_client()
+                # self.has_dask_client = True
+
+        if not evaluation_manager.local_data_saved:
+                    
+            print("\u2714 Preparing local data...            ", end="", flush=True)
+            print()
+            # self.memmap_map = load_local_data(evaluation_manager)
+            load_local_data(evaluation_manager)
+            # print('Completed!')
 
         root_dirpath = os.path.join(os.getcwd(), evaluation_manager.memmap_root_dirname)
         self.f = HMF.open_file(root_dirpath, mode='r+')
@@ -167,17 +182,21 @@ class EvaluationEngine():
         self.task_manager = TaskManager(
             **{k: v for k, v in evaluation_manager.__dict__.items() 
             if k in TASK_REQUIRED_KEYWORDS})
-        
-        if self.use_yarn_cluster:
+
+        if not evaluation_manager.local_data_saved:
             
-            print("\u2714 Uploading local data to S3 bucket...   ", end="", flush=True)
-            upload_local_data(self.task_manager)
-            print('Completed!')
+            if self.use_yarn_cluster:
+                
+                print("\u2714 Uploading local data to S3 bucket...   ", end="", flush=True)
+                upload_local_data(self.task_manager)
+                print('Completed!')
+                
+                print("\u2714 Preparing data on remote workers...   ", end="", flush=True)
+                self.dask_client.submit_per_node(download_local_data, self.task_manager)
+                print('Completed!')
+
+            evaluation_manager.local_data_saved = True
             
-            print("\u2714 Preparing data on remote workers...   ", end="", flush=True)
-            self.dask_client.submit_per_node(download_local_data, self.task_manager)
-            print('Completed!')
-        
         if debug_mode:
             print('\nRunning on debug mode!')
 
@@ -189,23 +208,23 @@ class EvaluationEngine():
 
             start_time = time.time()
 
-            total_splits = 0
-            for group_key in self.f.get_node_attr('/', key='sorted_group_keys'):
+            # total_splits = 0
+            # for group_key in self.f.get_node_attr('/', key='sorted_group_keys'):
 
-                if self.task_manager.orderby:
+            #     if self.task_manager.orderby:
 
-                    group_orderby_array = self.get_group_orderby_array(group_key)
+            #         group_orderby_array = self.get_group_orderby_array(group_key)
 
-                    cv = get_cv_splitter(
-                        self.task_manager.cross_validation_scheme, 
-                        self.task_manager.train_window, 
-                        self.task_manager.test_window,
-                        self.task_manager.min_train_window,
-                        group_orderby_array)
-                    total_splits += cv.get_n_splits()
+            #         cv = get_cv_splitter(
+            #             self.task_manager.cross_validation_scheme, 
+            #             self.task_manager.train_window, 
+            #             self.task_manager.test_window,
+            #             self.task_manager.min_train_window,
+            #             group_orderby_array)
+            #         total_splits += cv.get_n_splits()
 
 
-            print(total_splits)
+            # print(total_splits)
             print(time.time() - start_time)
 
 
@@ -275,15 +294,6 @@ class EvaluationEngine():
         os.chdir(evaluation_manager.initial_dirpath)
         
     def get_group_orderby_array(self, group_key):
-
-        
-
-
-        
-        # filepath = os.path.join(self.memmap_map['root_dirpath'], self.memmap_map['groups'][group_key]['arrays']['orderby_array']['filepath'])
-        # dtype = self.memmap_map['groups'][group_key]['arrays']['orderby_array']['dtype']
-        # shape = self.memmap_map['groups'][group_key]['arrays']['orderby_array']['shape']
-        # group_orderby_array = read_memmap(filepath, dtype, shape)
 
         group_orderby_array = self.f.get_array('/{}/orderby_array'.format(group_key))
         return group_orderby_array
@@ -434,6 +444,7 @@ class EvaluationEngine():
                 print('instance vcores: {}'.format(INSTANCE_TYPES[instance_type]['vCPU']))
                 print('instance memory: {} GB'.format(INSTANCE_TYPES[instance_type]['Mem']))
                 print('n_worker_nodes: {}'.format(self.n_worker_nodes))
+                print()
                 print('[ dask configurations ]')
                 print('local_client_n_workers: {}'.format(self.local_client_n_workers))
                 print('local_client_threads_per_worker: {}'.format(self.local_client_threads_per_worker))
@@ -455,16 +466,37 @@ class EvaluationEngine():
 
         self.taskq.join()
 
-        res = self.taskq.get_results()
-        res_pdf = pd.DataFrame(res, columns=['group_key', 'test_idx', 'eval_result', 'train_size', 'test_size', 'duration'])
+        res = copy.deepcopy(self.taskq.get_results())
+
+        tmp = self.data[[self.task_manager.orderby, constants.EF_ORDERBY_NAME]]
+        tmp.set_index(constants.EF_ORDERBY_NAME, inplace=True)
+        tmp_dict = tmp.to_dict()[self.task_manager.orderby]
+        tmp_dict = {k: str(v.date()) for k, v in tmp_dict.items()}
+
+        for idx, elem in enumerate(res):
+            res[idx][-2] = [tmp_dict[_elem] for _elem in elem[-2]]
+
+        res_pdf = pd.DataFrame(res, columns=['group_key', 'test_idx', 'eval_result', 'train_size', 'test_size', 'test_dates', 'duration'])
         return res_pdf.sort_values(by=['group_key', 'test_idx']).reset_index(drop=True)
+
+        
+        # res_pdf = pd.DataFrame(res, columns=['group_key', 'test_idx', 'eval_result', 'train_size', 'test_size', 'duration'])
+        # return res_pdf.sort_values(by=['group_key', 'test_idx']).reset_index(drop=True)
 
         return res
 
+    def get_evaluation_summary(self):
+
+        res = self.get_evaluation_results()
+
+        re_dict = {}
+        for group_key, grouped_pdf in res.groupby('group_key'):
+            re = np.sum(grouped_pdf['eval_result']*grouped_pdf['train_size'])/grouped_pdf['train_size'].sum()
+            re_dict[group_key] = re
+            
+        return pd.DataFrame([(k, v) for k, v in re_dict.items()], columns=['group_key', 'eval_result'])
+
     def get_prediction_results(self, group_key=None):
-
-
-        # print(os.getcwd())
 
         self.taskq.join()
